+++ config/routes.rb
@@ -92,6 +92,7 @@ Slice::Application.routes.draw do
 
     resources :exports do
       member do
+        get :file
         post :mark_unread
         post :progress
       end